Added some callbacks to dynamic channel
* Allow listener removal of channel during runtime
* Allow retrieval of channel name

+++ channels/drdynvc/client/drdynvc_main.h
@@ -45,11 +45,11 @@ struct _DVCMAN
 
 	drdynvcPlugin* drdynvc;
 
-	int num_plugins;
+	size_t num_plugins;
 	const char* plugin_names[MAX_PLUGINS];
 	IWTSPlugin* plugins[MAX_PLUGINS];
 
-	int num_listeners;
+	size_t num_listeners;
 	IWTSListener* listeners[MAX_PLUGINS];
 
 	wArrayList* channels;
